pulled correlator stuff out of geo
