+++ hera_mc/tests/test_geo_location.py
@@ -159,77 +159,6 @@ class TestGeo(TestHERAMC):
         stn = self.h.find_station_of_antenna(self.dna, 'now')
         self.assertTrue(stn == self.upte)
 
-    def test_ever_fully_connected(self):
-        now_list = self.h.get_all_fully_connected_at_date(at_date='now')
-        ever_list = self.h.get_all_fully_connected_ever()
-
-        self.assertEqual(len(now_list), 1)
-        self.assertEqual(len(ever_list), 2)
-
-        ever_ends = [loc['stop_date'] for loc in ever_list]
-        # The '==' notation below is required (rather than the pep8 suggestion of 'is') for the test to pass.
-        self.assertEqual(len(np.where(np.array(ever_ends) == None)[0]), len(now_list))
-        now_station_names = [loc['station_name'] for loc in now_list]
-        ever_station_names = [loc['station_name'] for loc in ever_list]
-        for name in now_station_names:
-            self.assertTrue(name in ever_station_names)
-
-        # check that every location fully connected now appears in fully connected ever
-        for loc_i, loc in enumerate(now_list):
-            this_station_list = [ever_list[i] for i in np.where(np.array(ever_station_names) == loc['station_name'])[0]]
-            this_station_starts = [loc['start_date'] for loc in this_station_list]
-            this_index = np.where(np.array(this_station_starts) == loc['start_date'])[0]
-
-            if len(this_index) == 1:
-                this_index = this_index[0]
-                print(loc)
-                print(this_station_list[this_index])
-                self.assertEqual(this_station_list[this_index], loc)
-
-    def test_correlator_info(self):
-        corr_dict = self.h.get_cminfo_correlator(cm_csv_path=mc.test_data_path)
-
-        ant_names = corr_dict['antenna_names']
-        self.assertEqual(len(ant_names), 1)
-
-        corr_inputs = corr_dict['correlator_inputs']
-
-        stn_types = corr_dict['station_types']
-
-        index = np.where(np.array(ant_names) == 'HH0')[0]
-        self.assertEqual(len(index), 1)
-        index = index[0]
-
-        self.assertEqual(stn_types[index], 'herahex')
-
-        self.assertEqual(corr_inputs[index], ('DF8B2', 'DF8B1'))
-
-        self.assertEqual([int(name.split('HH')[1]) for name in ant_names],
-                         corr_dict['antenna_numbers'])
-
-        self.assertEqual(set(corr_dict['antenna_numbers']),
-                         set([0]))
-
-        self.assertTrue(corr_dict['cm_version'] is not None)
-
-        # cm_version should be the same as the git hash of m&c for the test data
-        mc_dir = os.path.dirname(os.path.realpath(__file__))
-        mc_git_hash = subprocess.check_output(['git', '-C', mc_dir, 'rev-parse', 'HEAD'],
-                                              stderr=subprocess.STDOUT).strip()
-        self.assertEqual(corr_dict['cm_version'], mc_git_hash)
-
-        expected_keys = ['antenna_numbers', 'antenna_names', 'station_types',
-                         'correlator_inputs', 'antenna_utm_datum_vals',
-                         'antenna_utm_tiles', 'antenna_utm_eastings',
-                         'antenna_utm_northings', 'antenna_positions',
-                         'cm_version', 'cofa_lat', 'cofa_lon', 'cofa_alt']
-        self.assertEqual(set(corr_dict.keys()), set(expected_keys))
-
-        cofa = self.h.cofa()[0]
-        self.assertEqual(cofa.lat, corr_dict['cofa_lat'])
-        self.assertEqual(cofa.lon, corr_dict['cofa_lon'])
-        self.assertEqual(cofa.elevation, corr_dict['cofa_alt'])
-
 
 if __name__ == '__main__':
     unittest.main()
